feat(form-field): modularize components and enhance context management
