+++ packages/bleh-ui/src/components/form-field/components/form-field.error.tsx
@@ -1,0 +1,12 @@
+import { useContext, useEffect } from "react";
+import { FormFieldContext } from "../form-field";
+
+export const FormFieldError = ({ children }) => {
+  const [, setContext] = useContext(FormFieldContext);
+
+  useEffect(() => {
+    setContext((prevContext) => ({ ...prevContext, error: children }));
+  }, [children, setContext]);
+
+  return null;
+};
