+++ packages/bleh-ui/src/components/form-field/components/form-field.label.tsx
@@ -1,0 +1,12 @@
+import { useContext, useEffect } from "react";
+import { FormFieldContext } from "../form-field";
+
+export const FormFieldLabel = ({ children }) => {
+  const [, setContext] = useContext<object>(FormFieldContext);
+
+  useEffect(() => {
+    setContext((prevContext) => ({ ...prevContext, label: children }));
+  }, [children, setContext]);
+
+  return null;
+};
